math v0.0.1 Math

Mathematical functions and constants.

Types

x :: number

y :: number

Functions

x <~> y

Specs

number <~> number :: boolean

Equality-test for whether x and y are nearly equal.

This is useful when working with floating-point numbers, as these introduce small rounding errors.

Examples

iex> 2.3 - 0.3 == 2.0
false
iex> 2.3 - 0.3 <~> 2.0
true 

acos(x)

Specs

acos(x) :: float

Computes the arc cosine of x. (expressed in radians)

acosh(x)

Specs

acosh(x) :: float

Computes the inverse hyperbolic cosine of x.

asin(x)

Specs

asin(x) :: float

Computes the arc sine of x. (expressed in radians)

asinh(x)

Specs

asinh(x) :: float

Computes the inverse hyperbolic sine of x.

atan(x)

Specs

atan(x) :: float

Computes the arc tangent of x. (expressed in radians)

atan2(y, x)

Specs

atan2(y, x) :: float

Computes the arc tangent given y and x. (expressed in radians)

atanh(x)

Specs

atanh(x) :: float

Computes the inverse hyperbolic tangent of x.

cos(x)

Specs

cos(x) :: float

Computes the cosine of x.

x is interpreted as a value in radians.

cosh(x)

Specs

cosh(x) :: float

Computes the hyperbolic cosine of x (expressed in radians).

deg2rad(x)

Specs

deg2rad(x) :: float

Converts degrees to radians

Examples

iex>Math.deg2rad(180)
3.141592653589793

e()

Specs

e :: float

The mathematical constant ℯ (e).

exp(x)

Specs

exp(x) :: float

Returns ℯ to the xth power.

isqrt(x)

Returns the non-negative integer square root of x (rounded towards zero)

Does not accept negative numbers as input.

Examples

iex> Math.isqrt(100)
10
iex> Math.isqrt(16)
4
iex> Math.isqrt(65536)
256
iex> Math.isqrt(10)
3

log(x)

Specs

log(x) :: float

Returns the natural logarithm (base ℯ) of x.

See also e/0.

log10(x)

Specs

log10(x) :: float

Computes the common logarithm (base 10) of x.

log2(x)

Specs

log2(x) :: float

Returns the binary logarithm (base 2) of x.

pi()

Specs

pi :: float

The mathematical constant π (pi).

The ratio of a circle's circumference to its diameter. The returned number is a floating-point approximation (as π is irrational)

pow(x, n)

Specs

pow(number, number) :: number

Arithmetic exponentiation. Returns x to the n -th power.

When both x and n are integers and n is positive, returns an integer. When n is a negative integer, returns a float. When working with integers, the Exponentiation by Squaring algorithm is used, to allow for a fast and precise result.

When one of the numbers is a float, returns a float by using erlang's :math.pow/2 function.

It is possible to calculate roots by choosing n between 0.0 and 1.0 (To calculate the p -th-root, pass 1/p to the function)

Examples

iex> Math.pow(2, 4)
16
iex> Math.pow(2.0, 4)
16.0
iex> Math.pow(2, 4.0)
16.0
iex> Math.pow(5, 100)
7888609052210118054117285652827862296732064351090230047702789306640625
iex> Math.pow(5.0, 100)
7.888609052210118e69
iex> Math.pow(2, (1/2))
1.4142135623730951

rad2deg(x)

Specs

rad2deg(x) :: float

Converts radians to degrees

Examples

iex>Math.rad2deg(Math.pi)
180.0

sin(x)

Specs

sin(x) :: float

Computes the sine of x.

x is interpreted as a value in radians.

sinh(x)

Specs

sinh(x) :: float

Computes the hyperbolic sine of x (expressed in radians).

sqrt(x)

Specs

sqrt(x) :: float

Returns the non-negative square root of x.

tan(x)

Specs

tan(x) :: float

Computes the tangent of x (expressed in radians).

tanh(x)

Specs

tanh(x) :: float

Computes the hyperbolic tangent of x (expressed in radians).

tau()

Specs

tau :: float

The mathematical constant τ (tau).

The ratio of a circle's circumference to its radius. Defined as 2 * π, but preferred by some mathematicians. The returned number is a floating-point approximation (as τ is irrational)